TrillionDollarWords

Documentation for TrillionDollarWords.

TrillionDollarWords.load_all_data

TrillionDollarWords.load_all_sentences

TrillionDollarWords.load_cpi_data

TrillionDollarWords.load_market_data

TrillionDollarWords.load_ppi_data

TrillionDollarWords.load_training_sentences

TrillionDollarWords.load_ust_data

TrillionDollarWords.load_all_data — Method

load_all_data()

Load the combined dataset from the artifact. This dataset combines all sentences and the market data used in the paper.

The doc_id column is the unique identifier of the document.

The date column is the date of the event.

The event_type column is the type of event (meeting minutes, speech, or press conference).

The labels in label are predicted by the model proposed in the paper.

We use the RoBERTa-large model finetuned on the combined data to label all the filtered sentences in the meeting minutes, speeches, and press conferences.

The sentence column is the sentence itself.

The score column is the softmax probability of the label.

The speaker column is the speaker of the sentence (if applicable).

The value columns is the value of the market indicator (CPI, PPI, or UST).

The indicator column is the market indicator (CPI, PPI, or UST).

The maturity column is the maturity of the UST (if applicable).

TrillionDollarWords.load_all_sentences — Method

load_all_sentences()

Load the dataset with all sentences from the artifact. This is the complete dataset with sentences from press conferences, meeting minutes, and speeches.

The doc_id column is the unique identifier of the document.

The date column is the date of the event.

The event_type column is the type of event (meeting minutes, speech, or press conference).

The labels in label are predicted by the model proposed in the paper.

We use the RoBERTa-large model finetuned on the combined data to label all the filtered sentences in the meeting minutes, speeches, and press conferences.

The sentence column is the sentence itself.

The score column is the softmax probability of the label.

The speaker column is the speaker of the sentence (if applicable).

TrillionDollarWords.load_cpi_data — Method

load_cpi_data()

Load the CPI data from the artifact. This is the CPI data used in the paper.

The date column is the date of the event.

The value columns is the value of the market indicator (CPI, PPI, or UST).

The indicator column is the market indicator (CPI, PPI, or UST).

TrillionDollarWords.load_market_data — Method

load_market_data()

Load the combined market data from the artifact. This dataset combines the CPI, PPI and UST data used in the paper.

The date column is the date of the event.

The value columns is the value of the market indicator (CPI, PPI, or UST).

The indicator column is the market indicator (CPI, PPI, or UST).

The maturity column is the maturity of the UST (if applicable).

TrillionDollarWords.load_ppi_data — Method

load_ppi_data()

Load the PPI data from the artifact. This is the PPI data used in the paper.

The date column is the date of the event.

The value columns is the value of the market indicator (CPI, PPI, or UST).

The indicator column is the market indicator (CPI, PPI, or UST).

TrillionDollarWords.load_training_sentences — Method

load_training_sentences()

Load the dataset with training sentences from the artifact. This is a combined dataset containing sentences from press conferences, meeting minutes, and speeches.

The sentence column is the sentence itself.

The year column is the year of the event.

The labels in label are the manually annotated labels from the paper.

The seed column is the seed that was used to split the data into train and test set in the paper.

The sentence_splitting column indicates if the sentence was split or not (see the paper for details).

The event_type column is the type of event (meeting minutes, speech, or press conference).

The split column indicates if the sentence is in the train or test set.

TrillionDollarWords.load_ust_data — Method

load_ust_data()

Load the UST (treasury yields) data from the artifact. This is the UST data used in the paper.

The date column is the date of the event.

The value columns is the value of the market indicator (CPI, PPI, or UST).

The indicator column is the market indicator (CPI, PPI, or UST).

The maturity column is the maturity of the UST (if applicable).